// src/app/services/api/models/wb/similar/wb-similar-product.interface.ts
export interface WBSimilarProduct {
  brand: string;
  name: string;
  brandId: number;
  diffPrice: boolean;
  feedbacks: number;
  id: number;
  kindId: number;
  pics: number;
  priceU: number;
  rating: number;
  root: number;
  sale: number;
  salePriceU: number;
  siteBrandId: number;
  subjectId: number;
  subjectParentId: number;
}